--- README.old.md
@@ -1,1 +1,0 @@
-# mathemagic
